requirements.txt
torch>=1.9.0
torchvision>=0.10.0
streamlit>=1.22.0
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.21.0
matplotlib>=3.3.0
scikit-learn>=1.0.0
efficientnet-pytorch>=0.7.0
psutil>=5.8.0
segmentation-models-pytorch>=0.3.0
